requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-environ==0.11.2
django-ipware==7.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-ipware==3.0.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
resend==2.4.0
rpds-py==0.22.3
secure-smtplib==0.1.1
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
